requirements.txt
python-dotenv
aiogram==2.23.1
torch
transformers